Refactored io.circuitparser into subpackage with transparent interface to different implementations
