+++ pygsti/io/circuitparser/__init__.py
@@ -1,4 +1,4 @@
-""" A text parser for reading GST input files. """
+"""Implementations of a text parser for reading GST input files."""
 #***************************************************************************************************
 # Copyright 2015, 2019 National Technology & Engineering Solutions of Sandia, LLC (NTESS).
 # Under the terms of Contract DE-NA0003525 with NTESS, the U.S. Government retains certain rights
@@ -39,15 +39,31 @@ pstring   :: [ prep ] string
 ppstring  :: pstring [ povm ]
 """
 
+warn_msg = """
+An optimized Cython-based implementation of `{module}` is available as
+an extension, but couldn't be imported. This might happen if the
+extension has not been built. `pip install cython`, then reinstall
+pyGSTi to build Cython extensions. Alternatively, setting the
+environment variable `PYGSTI_NO_CYTHON_WARNING` will suppress this
+message.
+""".format(module=__name__)
+
+try:
+    # Import cython implementation if it's been built...
+    from .fastcircuitparser import *
+except ImportError:
+    # ... If not, fall back to the python implementation, with a warning.
+    import os as _os
+    import warnings as _warnings
+
+    if 'PYGSTI_NO_CYTHON_WARNING' not in _os.environ:
+        _warnings.warn(warn_msg)
+
+    from .slowcircuitparser import *
 
 import warnings as _warnings
 from ply import lex, yacc
-from ..objects import label as _lbl
-
-try:
-    from .fastcircuitparser import fast_parse_circuit as _fast_parse_circuit
-except ImportError:
-    _fast_parse_circuit = None
+from ...objects import label as _lbl
 
 
 class CircuitLexer:
@@ -191,8 +207,16 @@ class CircuitParser(object):
                                      tabmodule='pygsti.io.parsetab_string')
             self.parse = self.ply_parse
         else:
-            self._parser = SimpleCircuitParser()
-            self.parse = self._parser.parse
+            self.parse = self._parse
+
+    def _parse(self, code, create_subcircuits=True, integerize_sslbls=True):
+        """ Parse a circuit string
+
+        This method will dispatch to the optimized Cython
+        implementation, if available. Otherwise, the slower native
+        python implementation will be used.
+        """
+        return parse_circuit(code, create_subcircuits, integerize_sslbls)
 
     @property
     def lookup(self):
@@ -376,187 +400,3 @@ class CircuitParser(object):
         self._lexer.input(code)
         result = self._parser.parse(lexer=self._lexer)
         return result, labels
-
-
-class SimpleCircuitParser(object):
-    """ Parser for text-format operation sequences """
-
-    def __init__(self):
-        if _fast_parse_circuit is not None:
-            self.parse = self._fast_parse
-        else:
-            self.parse = self._slow_parse
-        self.warned = False
-
-    def _fast_parse(self, code, create_subcircuits=True, integerize_sslbls=True):
-        return _fast_parse_circuit(code, create_subcircuits, integerize_sslbls)
-
-    def _slow_parse(self, code, create_subcircuits=True, integerize_sslbls=True):
-
-        if not self.warned:
-            _warnings.warn(("Because you don't have Cython installed, pyGSTi is using"
-                            " a pure-Python parser.  This is fine, but it means loading"
-                            " text file will be somewhat slower."))
-            self.warned = True
-
-        if '@' in code:  # format:  <string>@<line_labels>
-            code, labels = code.split('@')
-            labels = labels.strip("( )")  # remove opening and closing parenthesis
-            def process(x): return int(x) if x.strip().isdigit() else x.strip()
-            labels = tuple(map(process, labels.split(',')))
-        else:
-            labels = None
-
-        result = []
-        code = code.replace('*', '')
-        i = 0; end = len(code); segment = 0
-
-        while(True):
-            if i == end: break
-            lbls_list, i, segment = self.get_next_lbls(code, i, end, create_subcircuits, integerize_sslbls, segment)
-            result.extend(lbls_list)
-
-        return result, labels
-
-    def get_next_lbls(self, s, start, end, create_subcircuits, integerize_sslbls, segment):
-
-        if s[start] == "(":
-            i = start + 1
-            lbls_list = []
-            while i < end and s[i] != ")":
-                lbls, i, segment = self.get_next_lbls(s, i, end, create_subcircuits, integerize_sslbls, segment)
-                lbls_list.extend(lbls)
-            if i == end: raise ValueError("mismatched parenthesis")
-            i += 1
-            exponent, i = self.parse_exponent(s, i, end)
-
-            if create_subcircuits:
-                if len(lbls_list) == 0:  # special case of {}^power => remain empty
-                    return [], i, segment
-                else:
-                    tmp = _lbl.Label(lbls_list)  # just for total sslbs - should probably do something faster
-                    return [_lbl.CircuitLabel('', lbls_list, tmp.sslbls, exponent)], i, segment
-            else:
-                return lbls_list * exponent, i, segment
-
-        elif s[start] == "[":  # layer
-            i = start + 1
-            lbls_list = []
-            while i < end and s[i] != "]":
-                lbls, i, segment = self.get_next_lbls(s, i, end, create_subcircuits, integerize_sslbls, segment)
-                lbls_list.extend(lbls)
-            if i == end: raise ValueError("mismatched parenthesis")
-            i += 1
-            exponent, i = self.parse_exponent(s, i, end)
-
-            if len(lbls_list) == 0:
-                to_exponentiate = _lbl.LabelTupTup(())
-            elif len(lbls_list) > 1:
-                time = max([l.time for l in lbls_list])
-                # create a layer label - a label of the labels within square brackets
-                to_exponentiate = _lbl.LabelTupTup(tuple(lbls_list), time)
-            else:
-                to_exponentiate = lbls_list[0]
-            return [to_exponentiate] * exponent, i, segment
-
-        else:
-            lbls, i, segment = self.get_next_simple_lbl(s, start, end, integerize_sslbls, segment)
-            exponent, i = self.parse_exponent(s, i, end)
-            return lbls * exponent, i, segment
-
-    def get_next_simple_lbl(self, s, start, end, integerize_sslbls, segment):
-        i = start
-        c = s[i]
-        if segment == 0 and s[i:i + 3] == 'rho':
-            i += 3; segment = 1
-        elif segment <= 1:
-            if (c == 'G' or c == 'I'):
-                i += 1; segment = 1
-            elif c == 'M':
-                i += 1; segment = 2  # marks end - no more labels allowed
-            elif c == '{':
-                i += 1
-                if i < end and s[i] == '}':
-                    i += 1
-                    return [], i, segment
-                else:
-                    raise ValueError("Invalid '{' at: %s..." % s[i - 1:i + 4])
-            else:
-                raise ValueError("Invalid prefix at: %s..." % s[i:i + 5])
-        else:
-            raise ValueError("Invalid prefix at: %s..." % s[i:i + 5])
-
-        while i < end:
-            c = s[i]
-            if 'a' <= c <= 'z' or '0' <= c <= '9' or c == '_':
-                i += 1
-            else:
-                break
-        name = s[start:i]; last = i
-
-        args = []
-        while i < end and s[i] == ';':
-            i += 1
-            last = i
-            while i < end:
-                c = s[i]
-                if 'a' <= c <= 'z' or '0' <= c <= '9' or c == '_' or c == 'Q' or c == '.' or c == '/':
-                    i += 1
-                else:
-                    break
-            args.append(s[last:i]); last = i
-
-        sslbls = []
-        while i < end and s[i] == ':':
-            i += 1
-            last = i
-            while i < end:
-                c = s[i]
-                if 'a' <= c <= 'z' or '0' <= c <= '9' or c == '_' or c == 'Q':
-                    i += 1
-                else:
-                    break
-            if integerize_sslbls:
-                try:
-                    val = int(s[last:i])
-                except:
-                    val = s[last:i]
-                sslbls.append(val); last = i
-            else:
-                sslbls.append(s[last:i]); last = i
-
-        if i < end and s[i] == '!':
-            i += 1
-            last = i
-            while i < end:
-                c = s[i]
-                if '0' <= c <= '9' or c == '.':
-                    i += 1
-                else:
-                    break
-            time = float(s[last:i])
-        else:
-            time = 0.0
-
-        if len(args) == 0:
-            if len(sslbls) == 0:
-                return [_lbl.LabelStr(name, time)], i, segment
-            else:
-                return [_lbl.LabelTup((name,) + tuple(sslbls), time)], i, segment
-        else:
-            return [_lbl.LabelTupWithArgs((name, 2 + len(args)) + tuple(args) + tuple(sslbls), time)], i, segment
-
-    def parse_exponent(self, s, i, end):
-        #z = re.match("\^([0-9]+)", s[i:end])
-        exponent = 1
-        if i < end and s[i] == '^':
-            i += 1
-            last = i
-            while i < end:
-                c = s[i]
-                if '0' <= c <= '9':
-                    i += 1
-                else:
-                    break
-            exponent = int(s[last:i])
-        return exponent, i
